Initialized new project
This is the Django project template. It contains all the required
stuff to start working with Django framework. README file describes
the process of the dependencies installation and provide guides on
repository usage, set up etc.

The PULL REQUEST template suitable for GitHub is also included and
configured.

This repository comes with pre-configured PostgreSQL and pgAdmin4
services within a Docker-compose configuration. pgadmin config
provides an initialized connection to postgres container.